test: keep Models witnesses within lint budgets
